Custom 2DII colour and fill scales

A custom discrete colour and fill scales with colours from 2DII palettes.

Usage

scale_colour_2dii(palette = c("2dii", "1in1000"), colour_groups = NULL, ...)

scale_fill_2dii(palette = c("2dii", "1in1000"), colour_groups = NULL, ...)

Arguments

palette

String with the name of the colour scale to be used. If not specified then the general 2dii scale is used

colour_groups

A vector containing groups variable to which colours are assigned. It is needed when the data assigned to colour aesthetic are not all contained in colour aliases of the palette.

...

Other parameters passed on to ggplot2::discrete_scale().

Value

An object of class "ScaleDiscrete".

Examples

library(ggplot2, warn.conflicts = FALSE)
library(r2dii.plot, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)

ggplot(sda %>% filter(emission_factor_metric == "projected")) +
 geom_line(aes(x = year, y = emission_factor_value, colour = sector)) +
 scale_colour_2dii()
#> Error in scale_colour_2dii(): could not find function "scale_colour_2dii"

ggplot(mpg) +
  geom_histogram(aes(cyl, fill = class), position = "dodge", bins = 5) +
  scale_fill_2dii()
#> Error in scale_fill_2dii(): could not find function "scale_fill_2dii"

ggplot(sda %>% filter(emission_factor_metric == "projected")) +
 geom_line(aes(x = year, y = emission_factor_value, colour = sector)) +
 scale_colour_2dii(palette = "1in1000", colour_groups = sda$sector)
#> Error in scale_colour_2dii(palette = "1in1000", colour_groups = sda$sector): could not find function "scale_colour_2dii"